update working code
