+++ src/core/capture/chrome-api-full-page.ts
@@ -1,0 +1,5 @@
+export default async function () {
+  const isFullPage = false;
+
+  return await chrome.tabs.captureVisibleTab();
+}
