+++ src/core/capture/chrome-api-visible-content.ts
@@ -1,0 +1,11 @@
+export default function () {
+  const isFullPage = false;
+  const capturePlace = 'background';
+  const captureType = 'visible';
+
+  return {
+    isFullPage: isFullPage,
+    capturePlace: capturePlace,
+    captureType: captureType,
+  };
+}
