+++ src/core/content/ext.css
@@ -1,0 +1,39 @@
+#scan2ai-html,
+#scan2ai-html * {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  cursor: crosshair;
+}
+
+#scan2ai-html canvas {
+  display: block;
+  width: auto;
+  height: auto;
+}
+
+#scan2ai-html .hidden {
+  display: none !important;
+}
+
+/* #scan2ai-html #backdrop {
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: #00000050;
+} */
+
+#scan2ai-html #scan2ai-select {
+  position: absolute;
+  background: #00000050;
+  z-index: 1000;
+}
+
+#scan2ai-html #scan2ai-result {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
