+++ src/core/content/ext.html
@@ -1,0 +1,18 @@
+<div id="backdrop"></div>
+<div id="scan2ai-select" class="hidden"></div>
+<img id="img" class="hidden">
+<div id="scan2ai-result" class="hidden">
+    <div>
+        Interact with your capture area
+        <span>x</span>
+    </div>
+    <div>
+        <div style="width: fit-content; border: 1px solid red;">
+            <canvas id="canvas"></canvas>
+        </div>
+    </div>
+    <div>
+        <div>Bot: ABC</div>
+        <div>You</div>
+    </div>
+</div>
