+++ src/core/selecting/drag-n-drop.ts
@@ -1,83 +1,124 @@
-let selectingStart = false;
+import type { Options } from '@/types/selecting';
 
-document.addEventListener('mousemove', onMouseUpdate, false);
-document.addEventListener('mouseenter', onMouseUpdate, false);
-document.addEventListener('mousedown', onMouseDown, false);
-document.addEventListener('mouseup', onMouseUp, false);
+export default function () {
+  let isSelecting: boolean = false;
+  let clientX: number | null = null;
+  let clientY: number | null = null;
+  let pageX: number | null = null;
+  let pageY: number | null = null;
+  let width: number | null = null;
+  let height: number | null = null;
 
-function onMouseUpdate(e) {
-  // if (selectingMode === 1) return;
-  if (!selectingStart) return;
-  let select = document.getElementById('scan2ai-select');
-  let currentX = e.pageX;
-  let currentY = e.pageY;
-  // width = Math.max(currentX, x) - Math.min(currentX, x);
-  // height = Math.max(currentY, y) - Math.min(currentY, y);
-  width = currentX - x;
-  height = currentY - y;
-  // select.style.top = e.clientY + 'px';
-  // select.style.left = e.clientX + 'px';
-  select.style.width = width + 'px';
-  select.style.height = height + 'px';
-}
-
-function onMouseDown(e) {
-  if (e.which !== 1) return;
-  console.log('onMouseDown', selectingStart);
-  selectingStart = true;
-  x = e.pageX;
-  y = e.pageY;
-  let select = document.getElementById('scan2ai-select');
-  select.classList.remove('hidden');
-  let result = document.getElementById('scan2ai-result');
-  result.classList.add('hidden');
-  setSelected(y, x);
-}
-
-function setSelected(top, left, width = 0, height = 0) {
-  let select = document.getElementById('scan2ai-select');
-  select.style.top = top + 'px';
-  select.style.left = left + 'px';
-  select.style.width = width + 'px';
-  select.style.height = height + 'px';
-}
-
-function onMouseUp(e) {
-  console.log('onMouseUp', selectingStart);
-  if (!selectingStart) return;
-  if (!width || !height) {
-    selectingStart = false;
-    return;
+  let options: Options = {
   }
-  selectingStart = false;
-  console.log('onMouseUp', selectingStart);
-  crop();
-  let select = document.getElementById('scan2ai-select');
-  select.classList.add('hidden');
-  let result = document.getElementById('scan2ai-result');
-  result.classList.remove('hidden');
-}
 
-function getBodyHeight() {
-  var body = document.body,
-    html = document.documentElement;
+  function clearSelection() {
+    clientX = null;
+    clientY = null;
+    pageX = null;
+    pageY = null;
+    width = null;
+    height = null;
+    if (!options.window) return;
+    const selection = options.window.getSelection()!;
+    // Remove selection to prevent selection bug
+    if (selection.empty) {
+      selection.empty();
+    } else if (selection.removeAllRanges) {  // Firefox
+      selection.removeAllRanges();
+    }
+  }
 
-  return Math.max(body.scrollHeight, body.offsetHeight, html.clientHeight, html.scrollHeight, html.offsetHeight);
-}
+  function initEventListeners(): void {
+    document.addEventListener('mousemove', onMouseUpdate, false);
+    document.addEventListener('mouseenter', onMouseUpdate, false);
+    document.addEventListener('mousedown', onMouseDown, false);
+    document.addEventListener('mouseup', onMouseUp, false);
+  }
+  function destroyEventListeners(): void {
+    document.removeEventListener('mousemove', onMouseUpdate);
+    document.removeEventListener('mouseenter', onMouseUpdate);
+    document.removeEventListener('mousedown', onMouseDown);
+    document.removeEventListener('mouseup', onMouseUp);
+  }
 
-export function init() {
-  let style = document.createElement('style');
-  style.textContent = extCss();
-  style.id = 'scan2ai-style';
-  document.body.appendChild(style);
-  let html = document.createElement('div');
-  html.id = 'scan2ai-html';
-  html.style.position = 'absolute';
-  html.style.top = '0';
-  html.style.left = '0';
-  html.style.width = '100%';
-  html.style.height = getBodyHeight() + 'px';
-  html.style.zIndex = '2147483640';
-  html.innerHTML = extHtml();
-  document.body.appendChild(html);
+  function onMouseDown(event: MouseEvent): void {
+    if (event.button !== 0) return;
+    isSelecting = true;
+    clientX = event.clientX;
+    clientY = event.clientY;
+    pageX = event.pageX;
+    pageY = event.pageY;
+    const select = document.getElementById('scan2ai-select')!;
+    select.classList.remove('hidden');
+    const result = document.getElementById('scan2ai-result')!;
+    result.classList.add('hidden');
+    setSelected(pageY, pageX);
+    options.onSelectingStart?.();
+  }
+
+  function onMouseUpdate(event: MouseEvent): void {
+    if (!isSelecting || !pageX || !pageY) return;
+    const select = document.getElementById('scan2ai-select')!;
+    const currentX = event.pageX;
+    const currentY = event.pageY;
+    // width = Math.max(currentX, x) - Math.min(currentX, x);
+    // height = Math.max(currentY, y) - Math.min(currentY, y);
+    width = currentX - pageX;
+    height = currentY - pageY;
+    // select.style.top = e.clientY + 'px';
+    // select.style.left = e.clientX + 'px';
+    select.style.width = width + 'px';
+    select.style.height = height + 'px';
+    options.onSelectingUpdate?.();
+  }
+
+  function setSelected(top = 0, left = 0, width = 0, height = 0) {
+    const select = document.getElementById('scan2ai-select')!;
+    select.style.top = top + 'px';
+    select.style.left = left + 'px';
+    select.style.width = width + 'px';
+    select.style.height = height + 'px';
+  }
+
+  function onMouseUp() {
+    if (!isSelecting) return;
+    if (!clientX || !clientY || !pageX || !pageY || !width || !height) {
+      isSelecting = false;
+      return;
+    }
+    isSelecting = false;
+    const select = document.getElementById('scan2ai-select')!;
+    select.classList.add('hidden');
+    const result = document.getElementById('scan2ai-result')!;
+    result.classList.remove('hidden');
+
+    options.onSelectingEnd?.({
+      clientX,
+      clientY,
+      pageX,
+      pageY,
+      width,
+      height,
+    });
+    clearSelection();
+  }
+
+  function init(overrideOptions: Options): void {
+    // chèn html, js, event listeners
+    options = {
+      ...options,
+      ...overrideOptions || {}
+    };
+    initEventListeners();
+  }
+
+  function destroy(): void {
+    destroyEventListeners();
+  }
+
+  return {
+    init,
+    destroy,
+  };
 }
