+++ src/core/selecting/index.ts
@@ -1,10 +1,13 @@
-import { init as dragAndDrop } from "./drag-n-drop";
+import dragAndDrop from "./drag-n-drop";
 import clickToStartAndEnd from "./click-to-start-n-end";
 
 /**
  * Mode 1: Click and hold
  * Mode 2: Click to start and click to end
  */
-export const DRAG_AND_DROP = dragAndDrop;
-
-export const CLICK_TO_START_AND_END = clickToStartAndEnd;
+export default function () {
+  return {
+    selectingDragAndDrop: dragAndDrop,
+    selectingClickToStartAndEnd: clickToStartAndEnd
+  }
+}
