+++ src/types/chromemessage.ts
@@ -1,0 +1,5 @@
+export interface ChromeMessageRequest {
+  action: string,
+  attributes?: any,
+  value?: any
+}
