+++ src/types/selecting.ts
@@ -3,5 +3,14 @@ export interface Options {
   document?: Document,
   onSelectingStart?: Function,
   onSelectingUpdate?: Function,
-  onSelectingEnd?: Function,
+  onSelectingEnd?: (result: onSelectingResult) => void,
+}
+
+export interface onSelectingResult {
+  clientX: number,
+  clientY: number,
+  pageX: number,
+  pageY: number,
+  width: number,
+  height: number,
 }
